+++ launchdarkly/resource_launchdarkly_team_member.go
@@ -6,6 +6,7 @@ import (
 	"net/http"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+
 	ldapi "github.com/launchdarkly/api-client-go"
 )
 
@@ -117,20 +118,9 @@ func resourceTeamMemberRead(d *schema.ResourceData, metaRaw interface{}) error {
 	_ = d.Set(LAST_NAME, member.LastName)
 	_ = d.Set(ROLE, member.Role)
 
-	// The LD api returns custom role IDs (not keys). Since we want to set custom_roles with keys, we need to look up their IDs
-	customRoleKeys := make([]string, 0, len(member.CustomRoles))
-	for _, customRoleID := range member.CustomRoles {
-		roleRaw, res, err := handleRateLimit(func() (interface{}, *http.Response, error) {
-			return client.ld.CustomRolesApi.GetCustomRole(client.ctx, customRoleID)
-		})
-		role := roleRaw.(ldapi.CustomRole)
-		if isStatusNotFound(res) {
-			return fmt.Errorf("failed to find custom role key for ID %q", customRoleID)
-		}
-		if err != nil {
-			return fmt.Errorf("failed to retrieve custom role key for role ID %q: %v", customRoleID, err)
-		}
-		customRoleKeys = append(customRoleKeys, role.Key)
+	customRoleKeys, err := customRoleIDsToKeys(client, member.CustomRoles)
+	if err != nil {
+		return err
 	}
 	err = d.Set(CUSTOM_ROLES, customRoleKeys)
 	if err != nil {
@@ -139,27 +129,57 @@ func resourceTeamMemberRead(d *schema.ResourceData, metaRaw interface{}) error {
 	return nil
 }
 
-func resourceTeamMemberUpdate(d *schema.ResourceData, metaRaw interface{}) error {
-	client := metaRaw.(*Client)
-	memberID := d.Id()
-	memberRole := d.Get(ROLE).(string)
-	customRoleKeys := d.Get(CUSTOM_ROLES).(*schema.Set).List()
+// The LD api returns custom role IDs (not keys). Since we want to set custom_roles with keys, we need to look up their IDs
+func customRoleIDsToKeys(client *Client, ids []string) ([]string, error) {
+	customRoleKeys := make([]string, 0, len(ids))
+	for _, customRoleID := range ids {
+		roleRaw, res, err := handleRateLimit(func() (interface{}, *http.Response, error) {
+			return client.ld.CustomRolesApi.GetCustomRole(client.ctx, customRoleID)
+		})
+		role := roleRaw.(ldapi.CustomRole)
+		if isStatusNotFound(res) {
+			return nil, fmt.Errorf("failed to find custom role key for ID %q", customRoleID)
+		}
+		if err != nil {
+			return nil, fmt.Errorf("failed to retrieve custom role key for role ID %q: %v", customRoleID, err)
+		}
+		customRoleKeys = append(customRoleKeys, role.Key)
+	}
+	return customRoleKeys, nil
+}
 
-	// Since the LD API expects custom role IDs, we need to look up each key to retrieve the ID
-	customRoleIds := make([]string, 0, len(customRoleKeys))
-	for _, rawKey := range customRoleKeys {
-		key := rawKey.(string)
+// Since the LD API expects custom role IDs, we need to look up each key to retrieve the ID
+func customRoleKeysToIDs(client *Client, keys []string) ([]string, error) {
+	customRoleIds := make([]string, 0, len(keys))
+	for _, key := range keys {
 		roleRaw, res, err := handleRateLimit(func() (interface{}, *http.Response, error) {
 			return client.ld.CustomRolesApi.GetCustomRole(client.ctx, key)
 		})
 		role := roleRaw.(ldapi.CustomRole)
 		if isStatusNotFound(res) {
-			return fmt.Errorf("failed to find custom ID for key %q", key)
+			return nil, fmt.Errorf("failed to find custom ID for key %q", key)
 		}
 		if err != nil {
-			return fmt.Errorf("failed to retrieve custom role ID for key %q: %v", key, err)
+			return nil, fmt.Errorf("failed to retrieve custom role ID for key %q: %v", key, err)
 		}
 		customRoleIds = append(customRoleIds, role.Id)
+	}
+	return customRoleIds, nil
+}
+
+func resourceTeamMemberUpdate(d *schema.ResourceData, metaRaw interface{}) error {
+	client := metaRaw.(*Client)
+	memberID := d.Id()
+	memberRole := d.Get(ROLE).(string)
+	customRolesRaw := d.Get(CUSTOM_ROLES).(*schema.Set).List()
+
+	customRoleKeys := make([]string, len(customRolesRaw))
+	for i, cr := range customRolesRaw {
+		customRoleKeys[i] = cr.(string)
+	}
+	customRoleIds, err := customRoleKeysToIDs(client, customRoleKeys)
+	if err != nil {
+		return err
 	}
 
 	patch := []ldapi.PatchOperation{
@@ -168,7 +188,7 @@ func resourceTeamMemberUpdate(d *schema.ResourceData, metaRaw interface{}) error
 		patchReplace("/customRoles", &customRoleIds),
 	}
 
-	_, _, err := handleRateLimit(func() (interface{}, *http.Response, error) {
+	_, _, err = handleRateLimit(func() (interface{}, *http.Response, error) {
 		return handleNoConflict(func() (interface{}, *http.Response, error) {
 			return client.ld.TeamMembersApi.PatchMember(client.ctx, memberID, patch)
 		})